refactor: extract inline rendering into InlineRendering module
- Extract ~560 lines of inline rendering methods from BaseRenderer into
  concerns/inline_rendering.rb (walk_ordered, render_mixed_inline, all
  inline element dispatch, semx handling, stem/math rendering)
- BaseRenderer drops from 1755 to 1199 lines
- Replace respond_to? in Theme.load with fixed SETTERS hash lookup
- Add Theme::VALID_KEYS and SETTERS constants
- Migrate BIPM cover page to Liquid template (_bipm_cover.html.liquid)
